Extract platform code
